% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSNPScale.R
\name{calculateLocation}
\alias{calculateLocation}
\title{Get the X and Y coordinates of a string coded location on a plot}
\usage{
calculateLocation(location = NULL, axisLimits)
}
\arguments{
\item{location}{A character string indicating a location on a plot}
}
\value{
Returns a numeric vector storing the coordinates determined by string location
}
\description{
Function used by \code{addSNPScale()}
}
\keyword{internal}
