% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanglePlot.R
\name{getTipCoordinates}
\alias{getTipCoordinates}
\title{Get the relative and absolute coordinates of tips on a plotted phylogeny}
\usage{
getTipCoordinates(tipLabels)
}
\arguments{
\item{tipLabels}{a vector of the tip labels of a plotted phylogeny, which can be retrieved from a "phylo" object using: \code{tree$tip.label}}
}
\description{
Function used by \code{tanglePlot()}
}
\keyword{internal}
