% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRAXML.R
\name{getTreeFileWithSupportValues}
\alias{getTreeFileWithSupportValues}
\title{Retrieves the tree file with bootstrap values from RAxML output}
\usage{
getTreeFileWithSupportValues(analysisName)
}
\arguments{
\item{analysisName}{A character string indicating the name of the analysis ran in RAXML}
}
\value{
Returns a phylo tree object
}
\description{
Function used by \code{runRAXML()}
}
\keyword{internal}
