% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanglePlot.R
\name{plotLineBetweenTips}
\alias{plotLineBetweenTips}
\title{Plot lines between sets of coordinates in the plotting window}
\usage{
plotLineBetweenTips(coordsA, coordsB, col = "black", lwd = 1,
  lty = 1, xOffset = 0)
}
\arguments{
\item{coordsA}{a list of coordinates on an existing plot}

\item{coordsB}{a list of coordinates on an existing plot}

\item{col}{colour of the lines connecting tips. Defaults to black}

\item{lwd}{width of the lines connecting tips. Defaults to 1}

\item{lty}{type of the lines connecting tips. Defaults to 1 (solid)}

\item{xOffset}{The distance on X axis to displace the start and ends of connecting lines. Defaults to 0}
}
\description{
Function used by \code{tanglePlot()}
}
\keyword{internal}
