% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanglePlot.R
\name{plotLinesBetweenTips}
\alias{plotLinesBetweenTips}
\title{Plot lines between coordinates of tips on phylogenies}
\usage{
plotLinesBetweenTips(tipLocationsA, tipLocationsB, col = "black",
  lwd = 1, lty = 1, onlyIfDifferent = TRUE, offsetProp = NULL)
}
\arguments{
\item{tipLocationsA}{a list of coordinates of tips on a phylogenetic tree plot}

\item{tipLocationsB}{a list of coordinates of tips on a phylogenetic tree plot}

\item{col}{colour of the lines connecting tips. Defaults to black}

\item{lwd}{width of the lines connecting tips. Defaults to 1}

\item{lty}{type of the lines connecting tips. Defaults to 1 (solid)}

\item{onlyIfDifferent}{a boolean variable, if \code{TRUE} only plots lines between tips that are in different locations on the phylogenies, otherwise lines between all tips are plotted. Defaults to \code{TRUE}}

\item{offsetProp}{offset value to displace starts of connecting lines from tips, set as the proportion of plotting window. Defaults to \code{NULL} (off)}
}
\description{
Function used by \code{tanglePlot()}
}
\keyword{internal}
