% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanglePlot.R
\name{tanglePlot}
\alias{tanglePlot}
\title{Plot the different nucleotides present in an alignment of nucleotide sequences}
\usage{
tanglePlot(
  tree1,
  tree2,
  onlyIfDifferent = TRUE,
  connectingLine.col = "black",
  connectingLine.lwd = 1,
  connectingLine.lty = 1,
  offsetProp = NULL,
  leftMargins = c(0, 0, 0, 1),
  rightMargins = c(0, 1, 0, 0),
  ...
)
}
\arguments{
\item{tree1}{an object of class "phylo" to compared to \code{tree2}}

\item{tree2}{an object of class "phylo" to compared to \code{tree1}}

\item{onlyIfDifferent}{a boolean variable, if \code{TRUE} only plots lines between tips that are in different locations on the phylogenies, otherwise lines between all tips are plotted. Defaults to \code{TRUE}}

\item{connectingLine.col}{colour of the lines connecting tips. Defaults to black}

\item{connectingLine.lwd}{width of the lines connecting tips. Defaults to 1}

\item{connectingLine.lty}{type of the lines connecting tips. Defaults to 1 (solid)}

\item{offsetProp}{offset value to displace starts of connecting lines from tips, set as the proportion of plotting window. Defaults to \code{NULL} (off)}

\item{leftMargins}{margin parameters for the left tree plot. Defaults to \code{c(0,0,0,1)} to leave small space on right}

\item{rightMargins}{margin parameters for the right tree plot. Defaults to \code{c(0,1,0,0)} to leave small space on left}
}
\description{
Function uses coloured polygons to illustrate which nucleotides are present in each sequence at each position
}
\examples{
# Set the seed
set.seed(76263)

# Create a random phylogeny
tree <- ape::rtree(n=50, rooted=TRUE)

# Rotate one of the nodes in the phylogeny - with tip labels it
rotated <- ape::rotate(tree, node=c("t35", "t34"))

# Compare the random and rotated phylogeny
tanglePlot(tree, rotated, connectingLine.col="red",
           connectingLine.lty=2, show.tip.label=TRUE, offsetProp=0.02)
}
\keyword{compare}
\keyword{plot}
\keyword{pylogeny}
\keyword{tangle}
\keyword{tree}
